refactor: move ajv formatter around
SL-2295
