+++ src/rulesets/__tests__/reader.unit.ts
@@ -7,9 +7,9 @@ import { Dictionary } from '@stoplight/types/basic';
 import { when } from 'jest-when';
 import { IRule } from '../..';
 import Lint from '../../cli/commands/lint';
-import { formatAjv } from '../../formatters/ajv';
 import { readParsable } from '../../fs/reader';
 import { IRulesetFile } from '../../types/ruleset';
+import { formatAjv } from '../ajv';
 import { resolvePath } from '../path';
 import { readRuleset } from '../reader';
 import { validateRuleset } from '../validation';
